requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
django-celery-beat==2.5.0
flower==2.0.1
python-decouple==3.8
django-cors-headers==4.3.1
drf-spectacular==0.26.5
gunicorn==21.2.0
django-filter==23.3
channels==4.0.0
channels-redis==4.1.0
websockets==11.0.3